Move `ssh_util` from `galaxy_test.base` to `galaxy.security`
so it can be used by both unit tests of the `galaxy-app` package and by integration tests
